STAC-14719: Refactor to fix tests about project structure
