+++ cmd/monitor.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 	"gitlab.com/stackvista/stackstate-cli2/cmd/monitor"
-	sts "gitlab.com/stackvista/stackstate-cli2/generated/stackstate_api"
 	"gitlab.com/stackvista/stackstate-cli2/internal/di"
 )
 
@@ -19,8 +18,8 @@ func MonitorCommand(cli *di.Deps) *cobra.Command {
 	cmd.AddCommand(monitor.MonitorRunCommand(cli))
 	cmd.AddCommand(monitor.MonitorStatusCommand(cli))
 	cmd.AddCommand(monitor.MonitorEditCommand(cli))
-	cmd.AddCommand(monitor.MonitorEnableDisableCommand(cli, "disable", "Disable a monitor", sts.MONITORSTATUSVALUE_DISABLED.Ptr(), "disabled"))
-	cmd.AddCommand(monitor.MonitorEnableDisableCommand(cli, "enable", "Enable a monitor", sts.MONITORSTATUSVALUE_ENABLED.Ptr(), "enabled"))
+	cmd.AddCommand(monitor.MonitorDisableCommand(cli))
+	cmd.AddCommand(monitor.MonitorEnableCommand(cli))
 
 	return cmd
 }
